// apps/blink-map/services/galoy/graphql/generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  AuthToken: { input: string; output: string; }
  CentAmount: { input: number; output: number; }
  ContactAlias: { input: string; output: string; }
  CountryCode: { input: string; output: string; }
  DisplayCurrency: { input: string; output: string; }
  EmailAddress: { input: string; output: string; }
  EmailRegistrationId: { input: string; output: string; }
  EndpointId: { input: string; output: string; }
  EndpointUrl: { input: string; output: string; }
  Feedback: { input: string; output: string; }
  Hex32Bytes: { input: string; output: string; }
  Language: { input: string; output: string; }
  LnPaymentPreImage: { input: string; output: string; }
  LnPaymentRequest: { input: string; output: string; }
  LnPaymentSecret: { input: string; output: string; }
  Memo: { input: string; output: string; }
  Minutes: { input: string; output: string; }
  NotificationCategory: { input: string; output: string; }
  OnChainAddress: { input: string; output: string; }
  OnChainTxHash: { input: string; output: string; }
  OneTimeAuthCode: { input: string; output: string; }
  PaymentHash: { input: string; output: string; }
  Phone: { input: string; output: string; }
  SafeInt: { input: number; output: number; }
  SatAmount: { input: number; output: number; }
  Seconds: { input: number; output: number; }
  SignedAmount: { input: number; output: number; }
  SignedDisplayMajorAmount: { input: string; output: string; }
  Timestamp: { input: number; output: number; }
  TotpCode: { input: string; output: string; }
  TotpRegistrationId: { input: string; output: string; }
  TotpSecret: { input: string; output: string; }
  Username: { input: string; output: string; }
  WalletId: { input: string; output: string; }
};
// ...
export type Coordinates = {
  readonly __typename: 'Coordinates';
  readonly latitude: Scalars['Float']['output'];
  readonly longitude: Scalars['Float']['output'];
};

export type Country = {
  readonly __typename: 'Country';
  readonly id: Scalars['CountryCode']['output'];
  readonly supportedAuthChannels: ReadonlyArray<PhoneCodeChannelType>;
};

export type Currency = {
  readonly __typename: 'Currency';
  readonly flag: Scalars['String']['output'];
  readonly fractionDigits: Scalars['Int']['output'];
  readonly id: Scalars['ID']['output'];
  readonly name: Scalars['String']['output'];
  readonly symbol: Scalars['String']['output'];
};
// ...
export type Globals = {
  readonly __typename: 'Globals';
  readonly buildInformation: BuildInformation;
  readonly feesInformation: FeesInformation;
  /** The domain name for lightning addresses accepted by this Galoy instance */
  readonly lightningAddressDomain: Scalars['String']['output'];
  readonly lightningAddressDomainAliases: ReadonlyArray<Scalars['String']['output']>;
  /** Which network (mainnet, testnet, regtest, signet) this instance is running on. */
  readonly network: Network;
  /**
   * A list of public keys for the running lightning nodes.
   * This can be used to know if an invoice belongs to one of our nodes.
   */
  readonly nodesIds: ReadonlyArray<Scalars['String']['output']>;
  /** A list of countries and their supported auth channels */
  readonly supportedCountries: ReadonlyArray<Country>;
};
// ...
export type LnInvoicePaymentStatusPayload = {
  readonly __typename: 'LnInvoicePaymentStatusPayload';
  readonly errors: ReadonlyArray<Error>;
  readonly status?: Maybe<InvoicePaymentStatus>;
};
// ...
export type MapInfo = {
  readonly __typename: 'MapInfo';
  readonly coordinates: Coordinates;
  readonly title: Scalars['String']['output'];
};

export type MapMarker = {
  readonly __typename: 'MapMarker';
  readonly mapInfo: MapInfo;
  readonly username: Scalars['Username']['output'];
};

export type MobileVersions = {
  readonly __typename: 'MobileVersions';
  readonly currentSupported: Scalars['Int']['output'];
  readonly minSupported: Scalars['Int']['output'];
  readonly platform: Scalars['String']['output'];
};
// ...
export type OnChainTxFee = {
  readonly __typename: 'OnChainTxFee';
  readonly amount: Scalars['SatAmount']['output'];
};
// ...
export type OnChainUsdTxFee = {
  readonly __typename: 'OnChainUsdTxFee';
  readonly amount: Scalars['CentAmount']['output'];
};
// ...
export type PricePoint = {
  readonly __typename: 'PricePoint';
  readonly price: Price;
  /** Unix timestamp (number of seconds elapsed since January 1, 1970 00:00:00 UTC) */
  readonly timestamp: Scalars['Timestamp']['output'];
};

/** A public view of a generic wallet which stores value in one of our supported currencies. */
export type PublicWallet = {
  readonly __typename: 'PublicWallet';
  readonly currency: WalletCurrency;
  readonly id: Scalars['ID']['output'];
  /** @deprecated Shifting property to 'currency' */
  readonly walletCurrency: WalletCurrency;
};

export type Query = {
  readonly __typename: 'Query';
  readonly accountDefaultWallet: PublicWallet;
  readonly btcPriceList?: Maybe<ReadonlyArray<Maybe<PricePoint>>>;
  readonly businessMapMarkers: ReadonlyArray<MapMarker>;
  readonly currencyList: ReadonlyArray<Currency>;
  readonly globals?: Maybe<Globals>;
  readonly lnInvoicePaymentStatus: LnInvoicePaymentStatusPayload;
  readonly me?: Maybe<User>;
  readonly mobileVersions?: Maybe<ReadonlyArray<Maybe<MobileVersions>>>;
  readonly onChainTxFee: OnChainTxFee;
  readonly onChainUsdTxFee: OnChainUsdTxFee;
  readonly onChainUsdTxFeeAsBtcDenominated: OnChainUsdTxFee;
  /** Returns 1 Sat and 1 Usd Cent price for the given currency */
  readonly realtimePrice: RealtimePrice;
  /** @deprecated will be migrated to AccountDefaultWalletId */
  readonly userDefaultWalletId: Scalars['WalletId']['output'];
  readonly usernameAvailable?: Maybe<Scalars['Boolean']['output']>;
};
// ...
export type RealtimePrice = {
  readonly __typename: 'RealtimePrice';
  readonly btcSatPrice: PriceOfOneSatInMinorUnit;
  readonly denominatorCurrency: Scalars['DisplayCurrency']['output'];
  readonly id: Scalars['ID']['output'];
  /** Unix timestamp (number of seconds elapsed since January 1, 1970 00:00:00 UTC) */
  readonly timestamp: Scalars['Timestamp']['output'];
  readonly usdCentPrice: PriceOfOneUsdCentInMinorUnit;
};
// ...
export type User = {
  readonly __typename: 'User';
  readonly apiKeys: ReadonlyArray<ApiKey>;
  /**
   * Get single contact details.
   * Can include the transactions associated with the contact.
   * @deprecated will be moved to Accounts
   */
  readonly contactByUsername: UserContact;
  /**
   * Get full list of contacts.
   * Can include the transactions associated with each contact.
   * @deprecated will be moved to account
   */
  readonly contacts: ReadonlyArray<UserContact>;
  readonly createdAt: Scalars['Timestamp']['output'];
  readonly defaultAccount: Account;
  /** Email address */
  readonly email?: Maybe<Email>;
  readonly id: Scalars['ID']['output'];
  /**
   * Preferred language for user.
   * When value is 'default' the intent is to use preferred language from OS settings.
   */
  readonly language: Scalars['Language']['output'];
  readonly notificationSettings: UserNotificationSettings;
  /** Phone number with international calling code. */
  readonly phone?: Maybe<Scalars['Phone']['output']>;
  /** Whether TOTP is enabled for this user. */
  readonly totpEnabled: Scalars['Boolean']['output'];
  /**
   * Optional immutable user friendly identifier.
   * @deprecated will be moved to @Handle in Account and Wallet
   */
  readonly username?: Maybe<Scalars['Username']['output']>;
};
// ...
export type BusinessMapMarkersQuery = { readonly __typename: 'Query', readonly businessMapMarkers: ReadonlyArray<{ readonly __typename: 'MapMarker', readonly username: string, readonly mapInfo: { readonly __typename: 'MapInfo', readonly title: string, readonly coordinates: { readonly __typename: 'Coordinates', readonly latitude: number, readonly longitude: number } } }> };